requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pymongo>=4.4.0
neo4j>=5.10.0
numpy>=1.24.0
scikit-learn>=1.3.0
httpx>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
py-algorand-sdk>=2.0.0
pyteal>=0.20.0
